LyricSync: bug fixes + code simplifications
Bug fixes:
- MediaSessionTracker: volatile currentState, null-guard on session change,
  interpolate position during STATE_BUFFERING (fixes frozen/leading lyrics).
- BootReceiver: use correct shared prefs name ('lyricsync').
- LyricsProviderManager: safe shutdownNow + double-shutdown guard.
- SpicyLyricsProvider: skip events with negative/missing timestamps.
- InterludeDotView: revert container spring damping (5.0, 0.7) for correct
  appear/disappear animation.

Simplifications / cleanup:
- SyllableHighlighter: background-vocal words use setBackgroundMode(true);
  extract createBackgroundWords helper (removes duplicated per-word/per-line
  bg construction); drop dead ACTIVE/INACTIVE size multipliers and always-true
  layer-type guards.
- InterludeDotView: collapse redundant isPast early-return.
- AppLog: replace ThreadLocal SimpleDateFormat with a static instance.
- Extract SeekBars.bind() helper used by MainActivity and
  FloatingOverlayService (removes repeated listener boilerplate).
- FloatingOverlayService: animate interlude dots every frame (also while paused).

+++ app/src/main/java/com/lyricsync/app/renderer/Spline.java
@@ -116,7 +116,7 @@ public class Spline {
         m.set(l, p);
     }
 
-    private List<List<Double>> zerosMat(double r, int c) {
+    private List<List<Double>> zerosMat(int r, int c) {
         List<List<Double>> a = new ArrayList<>();
         for (int i = 0; i < r; i++) {
             List<Double> list = new ArrayList<>();
